+++ src/sphobjinv/error.py
@@ -10,7 +10,7 @@ Sphinx |objects.inv| files.
     5 Nov 2017
 
 **Copyright**
-    \(c) Brian Skinn 2016-2019
+    \(c) Brian Skinn 2016-2020
 
 **Source Repository**
     http://www.github.com/bskinn/sphobjinv
@@ -37,7 +37,3 @@ class VersionError(SphobjinvError):
     |objects.inv| files (see :doc:`here </syntax>`).
 
     """
-
-
-if __name__ == "__main__":  # pragma: no cover
-    print("Module not executable.")