+++ src/sphobjinv/inventory.py
@@ -10,7 +10,7 @@ Sphinx |objects.inv| files.
     7 Dec 2017
 
 **Copyright**
-    \(c) Brian Skinn 2016-2019
+    \(c) Brian Skinn 2016-2020
 
 **Source Repository**
     http://www.github.com/bskinn/sphobjinv
@@ -25,85 +25,33 @@ Sphinx |objects.inv| files.
 
 """
 
-from enum import Enum
+import re
+import ssl
+import urllib.request as urlrq
+import warnings
+from zlib import error as zlib_error
 
 import attr
+import certifi
+import jsonschema
+from jsonschema.exceptions import ValidationError
 
-from .data import DataObjStr
-
-
-class HeaderFields(Enum):
-    """|Enum| for various inventory-level data items.
-
-    A subset of these |Enum| values is used in various Regex,
-    JSON, and string formatting contexts within :class:`Inventory`
-    and :data:`schema.json_schema <sphobjinv.schema.json_schema>`.
-
-    """
-
-    #: Project name associated with an inventory
-    Project = "project"
-
-    #: Project version associated with an inventory
-    Version = "version"
-
-    #: Number of objects contained in the inventory
-    Count = "count"
-
-    #: The |str| value of this |Enum| member is accepted as a root-level
-    #: key in a |dict| to be imported into an :class:`Inventory`.
-    #: The corresponding value in the |dict| may contain any arbitrary data.
-    #: Its possible presence is accounted for in
-    #: :data:`schema.json_schema <sphobjinv.schema.json_schema>`.
-    #:
-    #: The data associated with this key are **ignored**
-    #: during import into an :class:`Inventory`.
-    Metadata = "metadata"
+from sphobjinv.data import _utf8_encode, DataObjStr
+from sphobjinv.enum import HeaderFields, SourceTypes
+from sphobjinv.fileops import readbytes
+from sphobjinv.re import pb_data, pb_project, pb_version
+from sphobjinv.schema import json_schema
+from sphobjinv.zlib import decompress
 
 
-class SourceTypes(Enum):
-    """|Enum| for the import mode used in instantiating an |Inventory|.
-
-    Since |Enum| keys iterate in definition order, the
-    definition order here defines the order in which |Inventory|
-    objects attempt to parse a source object passed to
-    :class:`Inventory.__init__() <Inventory>` either as a positional argument
-    or via the generic `source` keyword argument.
-
-    This order **DIFFERS** from the documentation order, which is
-    alphabetical.
-
-    """
-
-    #: No source; |Inventory| was instantiated with
-    #: :data:`~Inventory.project` and :data:`~Inventory.version`
-    #: as empty strings and
-    #: :data:`~Inventory.objects` as an empty |list|.
-    Manual = "manual"
-
-    #: Instantiation from a plaintext |objects.inv| |bytes|.
-    BytesPlaintext = "bytes_plain"
-
-    #: Instantiation from a zlib-compressed
-    #: |objects.inv| |bytes|.
-    BytesZlib = "bytes_zlib"
-
-    #: Instantiation from a plaintext |objects.inv| file on disk.
-    FnamePlaintext = "fname_plain"
-
-    #: Instantiation from a zlib-compressed |objects.inv| file on disk.
-    FnameZlib = "fname_zlib"
-
-    #: Instantiation from a |dict| validated against
-    #: :data:`schema.json_schema <sphobjinv.schema.json_schema>`.
-    DictJSON = "dict_json"
-
-    #: Instantiation from a zlib-compressed |objects.inv| file
-    #: downloaded from a URL.
-    URL = "url"
+# Cope with 'cmp' argument deprecation in attrs 19.2
+try:
+    _attr_wrapper = attr.s(slots=True, eq=False)
+except TypeError:  # pragma: no cover
+    _attr_wrapper = attr.s(slots=True, cmp=False)
 
 
-@attr.s(slots=True, cmp=False)
+@_attr_wrapper
 class Inventory(object):
     r"""Entire contents of an |objects.inv| inventory.
 
@@ -225,7 +173,7 @@ class Inventory(object):
     #: of the elements is anything other than |DataObjStr|.
     objects = attr.ib(init=False, default=attr.Factory(list), repr=False)
 
-    #: :class:`SourceTypes` |Enum| value indicating the type of
+    #: :class:`~sphobjinv.enum.SourceTypes` |Enum| value indicating the type of
     #: source from which the instance was generated.
     source_type = attr.ib(init=False, default=None)
 
@@ -241,6 +189,9 @@ class Inventory(object):
 
     #: zlib compression line for v2 |objects.inv| header
     header_zlib = "# The remainder of this file is compressed using zlib."
+
+    # Private class member for SSL context, since context creation is slow(?)
+    _sslcontext = ssl.create_default_context(cafile=certifi.where())
 
     @property
     def count(self):
@@ -333,7 +284,7 @@ class Inventory(object):
             If both `expand` and `contract` are |True|
 
         """
-        return list(_.as_rst for _ in self.objects)
+        return [_.as_rst for _ in self.objects]
 
     def __str__(self):  # pragma: no cover
         """Return concise, readable description of contents."""
@@ -348,8 +299,6 @@ class Inventory(object):
 
     def __attrs_post_init__(self):
         """Construct the inventory from the indicated source."""
-        from .data import _utf8_encode
-
         # List of sources
         src_list = (
             self._source,
@@ -460,26 +409,19 @@ class Inventory(object):
         # Extra empty string at the end puts a newline at the end
         # of the generated string, consistent with files
         # generated by Sphinx.
-
-        # Can't *-expand as a mixed argument in python 3.4, so have
-        # to assemble the strings sequentially
-        from itertools import chain
-
-        striter = chain(
-            [
+        return "\n".join(
+            (
                 self.header_preamble,
                 self.header_project.format(project=self.project),
                 self.header_version.format(version=self.version),
                 self.header_zlib,
-            ],
-            (
-                obj.data_line(expand=expand, contract=contract)
-                for obj in self.objects
-            ),
-            [""],
-        )  # DO want trailing newline
-
-        return "\n".join(striter).encode("utf-8")
+                *(
+                    obj.data_line(expand=expand, contract=contract)
+                    for obj in self.objects
+                ),
+                "",
+            )
+        ).encode("utf-8")
 
     def suggest(self, name, *, thresh=50, with_index=False, with_score=False):
         r"""Suggest objects in the inventory to match a name.
@@ -550,9 +492,6 @@ class Inventory(object):
             |cour|\ (as_rst, score, index)\ |/cour|
 
         """
-        import re
-        import warnings
-
         # Suppress any UserWarning about the speed issue
         with warnings.catch_warnings():
             warnings.simplefilter("ignore")
@@ -560,46 +499,40 @@ class Inventory(object):
 
         # Must propagate list index to include in output
         # Search vals are rst prepended with list index
-        srch_list = list(
-            "{0} {1}".format(i, o) for i, o in enumerate(self.objects_rst)
-        )
+        srch_list = ["{0} {1}".format(i, o) for i, o in enumerate(self.objects_rst)]
 
         # Composite each string result extracted by fuzzywuzzy
         # and its match score into a single string. The match
         # and score are returned together in a tuple.
-        results = list(
+        results = [
             "{0} {1}".format(*_)
             for _ in fwp.extract(name, srch_list, limit=None)
             if _[1] >= thresh
-        )
+        ]
 
         # Define regex for splitting the three components, and
         # use it to convert composite result string to tuple:
-        # (rst, score, index)
+        # result --> (rst, score, index)
         p_idx = re.compile("^(\\d+)\\s+(.+?)\\s+(\\d+)$")
-        results = list(
+        results = [
             (m.group(2), int(m.group(3)), int(m.group(1)))
             for m in map(p_idx.match, results)
-        )
+        ]
 
         # Return based on flags
         if with_score:
             if with_index:
                 return results
             else:
-                return list(tup[:2] for tup in results)
+                return [tup[:2] for tup in results]
         else:
             if with_index:
-                return list(tup[::2] for tup in results)
+                return [tup[::2] for tup in results]
             else:
-                return list(tup[0] for tup in results)
+                return [tup[0] for tup in results]
 
     def _general_import(self):
         """Attempt sequence of all imports."""
-        from zlib import error as ZlibError
-
-        from jsonschema.exceptions import ValidationError
-
         # Lookups for method names and expected import-failure errors
         importers = {
             SourceTypes.BytesPlaintext: self._import_plaintext_bytes,
@@ -610,9 +543,9 @@ class Inventory(object):
         }
         import_errors = {
             SourceTypes.BytesPlaintext: TypeError,
-            SourceTypes.BytesZlib: (ZlibError, TypeError),
+            SourceTypes.BytesZlib: (zlib_error, TypeError),
             SourceTypes.FnamePlaintext: (OSError, TypeError),
-            SourceTypes.FnameZlib: (OSError, TypeError, ZlibError),
+            SourceTypes.FnameZlib: (OSError, TypeError, zlib_error),
             SourceTypes.DictJSON: (ValidationError),
         }
 
@@ -623,9 +556,7 @@ class Inventory(object):
                 # No action for source types w/o a handler function defined.
                 continue
 
-            if self._try_import(
-                importers[st], self._source, import_errors[st]
-            ):
+            if self._try_import(importers[st], self._source, import_errors[st]):
                 self.source_type = st
                 return
 
@@ -651,8 +582,6 @@ class Inventory(object):
 
     def _import_plaintext_bytes(self, b_str):
         """Import an inventory from plaintext bytes."""
-        from .re import pb_data, pb_project, pb_version
-
         b_res = pb_project.search(b_str).group(HeaderFields.Project.value)
         project = b_res.decode("utf-8")
 
@@ -673,8 +602,6 @@ class Inventory(object):
 
     def _import_zlib_bytes(self, b_str):
         """Import a zlib-compressed inventory."""
-        from .zlib import decompress
-
         b_plain = decompress(b_str)
         p, v, o = self._import_plaintext_bytes(b_plain)
 
@@ -682,29 +609,22 @@ class Inventory(object):
 
     def _import_plaintext_fname(self, fn):
         """Import a plaintext inventory file."""
-        from .fileops import readbytes
-
         b_plain = readbytes(fn)
 
         return self._import_plaintext_bytes(b_plain)
 
     def _import_zlib_fname(self, fn):
         """Import a zlib-compressed inventory file."""
-        from .fileops import readbytes
-
         b_zlib = readbytes(fn)
 
         return self._import_zlib_bytes(b_zlib)
 
     def _import_url(self, url):
         """Import a file from a remote URL."""
-        import urllib.request as urlrq
-
-        import certifi
-
         # Caller's responsibility to ensure URL points
         # someplace safe/sane!
-        resp = urlrq.urlopen(url, cafile=certifi.where())
+        req = urlrq.Request(url, headers={"User-Agent": "Magic Browser"})
+        resp = urlrq.urlopen(req, context=self._sslcontext)  # noqa: S310
         b_str = resp.read()
 
         # Plaintext URL D/L is unreliable; zlib only
@@ -712,11 +632,6 @@ class Inventory(object):
 
     def _import_json_dict(self, d):
         """Import flat-dict composited data."""
-        import jsonschema
-
-        from .data import DataObjStr
-        from .schema import json_schema
-
         # Validate the dict against the schema. Schema
         # WILL allow an inventory with no objects here
         val = jsonschema.Draft4Validator(json_schema)
@@ -749,7 +664,7 @@ class Inventory(object):
 
         # Complain if remaining objects are anything other than the
         # valid inventory-level header keys
-        hf_values = set(e.value for e in HeaderFields)
+        hf_values = {e.value for e in HeaderFields}
         check_value = self._count_error and set(d.keys()).difference(hf_values)
         if check_value:
             # A truthy value here will be the contents
@@ -759,7 +674,3 @@ class Inventory(object):
 
         # Should be good to return
         return project, version, objects
-
-
-if __name__ == "__main__":  # pragma: no cover
-    print("Module not executable.")